requirements.txt
# Bot
discord.py>=2.4.0  # Python 3.13+ support (removed audioop dependency)
python-dotenv==1.0.0

# Database
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary>=2.9.11  # PostgreSQL support (binary wheels for all platforms)

# API & Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
jinja2==3.1.3  # Template engine
python-jose[cryptography]==3.3.0  # For JWT
python-multipart==0.0.6  # For form data
httpx==0.25.2  # For async HTTP (Discord OAuth)
fastapi-htmx==0.4.0  # HTMX integration helpers

# Utilities
pydantic-settings==2.1.0
structlog==24.1.0  # Better logging
aiofiles==23.2.1  # For async static file serving
itsdangerous==2.1.2  # For secure sessions
cryptography==41.0.7  # For encryption

# Optional (for enhanced features)
redis==4.6.0  # For caching and session storage (compatible with fastapi-limiter)
fastapi-limiter==0.1.5  # Rate limiting

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9
psutil==5.9.6  # For process management